renamed z_hookset.c/h and some cosmetics
